Dedupe eventfd test programs
