+++ cannon/testdata/common/go.mod
@@ -1,0 +1,5 @@
+module common
+
+go 1.23
+
+toolchain go1.23.8
