+++ cannon/testdata/go-1-23/syscall-eventfd/go.mod
@@ -3,3 +3,7 @@ module hello
 go 1.23.0
 
 toolchain go1.23.8
+
+require common v0.0.0
+
+replace common => ./../../common
