+++ cannon/testdata/go-1-24/syscall-eventfd/go.mod
@@ -3,3 +3,7 @@ module hello
 go 1.24
 
 toolchain go1.24.2
+
+require common v0.0.0
+
+replace common => ./../../common
